refactor: move util to own package
