+++ internal/service/util/util_test.go
@@ -1,4 +1,4 @@
-package service
+package util
 
 import (
 	"testing"
